BackEnd/requirements.txt
flask
python-dotenv
requests
PyJWT
mysql-connector-python